backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
bcrypt==4.0.1
SQLAlchemy==2.0.21
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.5.3
python-jose==3.3.0
requests==2.31.0
authlib==1.2.1
Werkzeug==2.3.7
psycopg2-binary==2.9.10 